switch box86 rpi4 package to box86-rpi4arm64 (#6)
all packages provide box86 so any other package can still depend on box86. previously the user could not guarantee the installation of the rpi4 box86 package, installing box86 would result in a random selection of any of the packages that provide it. now the user or script can select the box86-rpi4arm64 target explicitly